replace backend with typescript express app
