--- .secrets.env.example
@@ -1,3 +1,0 @@
-SUPERTOKENS_URI="find on https://supertokens.com/dashboard-saas"
-SUPERTOKENS_API_KEY="find on https://supertokens.com/dashboard-saas"
-SUPERTOKENS_DASHBOARD_KEY="you can make this anything"
