+++ backend/.env
@@ -1,0 +1,16 @@
+BACKEND_PORT=8080
+FRONTEND_PORT=3000
+
+DEV_DB_HOST=postgres_dev
+DEV_DB_DRIVER=postgres
+DEV_DB_USER=dev_user
+DEV_DB_PASSWORD=password
+DEV_DB_NAME=income_project_dev
+DEV_DB_PORT=5432
+
+TEST_DB_HOST=postgres_test
+TEST_DB_DRIVER=postgres
+TEST_DB_USER=test_user
+TEST_DB_PASSWORD=password
+TEST_DB_NAME=income_project_test
+TEST_DB_PORT=5433
