--- backend/.gitignore
@@ -1,3 +1,0 @@
-.DS_Store
-build/*
-log.txt
