+++ backend/.secrets.env.example
@@ -1,0 +1,2 @@
+SUPERTOKENS_URI="find on https://supertokens.com/dashboard-saas"
+SUPERTOKENS_API_KEY="find on https://supertokens.com/dashboard-saas"
