+++ backend/__tests__/app.test.ts
@@ -1,0 +1,10 @@
+import request from "supertest";
+
+import app from "../app";
+
+describe("Test app.ts", () => {
+  test("Catch-all route", async () => {
+    const res = await request(app).get("/");
+    expect(res.body).toEqual({ message: "Allo! Catch-all route." });
+  });
+});
