+++ backend/common/constants.ts
@@ -1,0 +1,10 @@
+import dotenv from "dotenv";
+
+dotenv.config();
+dotenv.config({ path: ".secrets.env" });
+
+export const backendPort = process.env.BACKEND_PORT;
+export const frontendPort = process.env.FRONTEND_PORT;
+
+export const supertokensURI = process.env.SUPERTOKENS_URI || "";
+export const supertokensAPIKey = process.env.SUPERTOKENS_API_KEY || "";
