--- backend/main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"recipient/config"
-	"recipient/models"
-	"recipient/server"
-)
-
-func main() {
-	config.Init()
-	models.ConnectDB()
-	server.Init()
-}
