+++ backend/server.ts
@@ -1,0 +1,6 @@
+import app from "./app";
+import { backendPort } from "./common/constants";
+
+app.listen(backendPort, () => {
+  console.log(`Server is running at http://localhost:${backendPort}`);
+});
