+++ backend/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "exclude": ["./coverage", "./build", "__tests__", "jest.config.js"],
+  "compilerOptions": {
+    "target": "es2016",                               /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
+    "module": "commonjs",                             /* Specify what module code is generated. */
+    "outDir": "./build",                              /* Specify an output folder for all emitted files. */
+    "esModuleInterop": true,                          /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
+    "forceConsistentCasingInFileNames": true,         /* Ensure that casing is correct in imports. */
+    "strict": true,                                   /* Enable all strict type-checking options. */
+    "noFallthroughCasesInSwitch": true,               /* Enable error reporting for fallthrough cases in switch statements. */
+    "skipLibCheck": true                              /* Skip type checking all .d.ts files. */
+  }
+}
