+++ backend_old/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"recipient/config"
+	"recipient/models"
+	"recipient/server"
+)
+
+func main() {
+	config.Init()
+	models.ConnectDB()
+	server.Init()
+}
