--- docker-compose.test.yaml
@@ -1,39 +1,0 @@
-services:
-  backend_test:
-    container_name: income_project_backend_test
-    env_file:
-      - .env
-      - .secrets.env
-    build:
-      context: ./
-      dockerfile: backend/Dockerfile.test
-    ports:
-      - 8080:8080
-    volumes:
-      - ./backend:/backend
-    depends_on:
-      - postgres_test
-    networks:
-      - income_project_test
-
-  postgres_test:
-    image: postgres:latest
-    container_name: postgres_test_container
-    environment:
-      - POSTGRES_USER=${TEST_DB_USER}
-      - POSTGRES_PASSWORD=${TEST_DB_PASSWORD}
-      - POSTGRES_DB=${TEST_DB_NAME}
-      - DATABASE_HOST=${TEST_DB_HOST}
-    ports:
-      - '4568:5432'
-    volumes:
-      - ./pg_data_test:/var/lib/postgresql/data
-    networks:
-      - income_project_test
-
-volumes:
-  pg_data_test:
-
-networks:
-  income_project_test:
-    driver: bridge
